dnos* -> dellos* (Rename module & updated copyright)  (#4888)
* Renamed the Modules from dnos* -> dellos*, updated copyright, corrected doc issues

* Removed the unwanted module import
