+++ network/dellos9/dellos9_command.py
@@ -1,5 +1,9 @@
 #!/usr/bin/python
 #
+# (c) 2015 Peter Sprygada, <psprygada@ansible.com>
+#
+# Copyright (c) 2016 Dell Inc.
+#
 # This file is part of Ansible
 #
 # Ansible is free software: you can redistribute it and/or modify
@@ -18,7 +22,7 @@
 
 DOCUMENTATION = """
 ---
-module: dnos9_command
+module: dellos9_command
 version_added: "2.2"
 short_description: Run commands on remote devices running Dell OS9
 description:
@@ -27,12 +31,12 @@ description:
     argument that will cause the module to wait for a specific condition
     before returning or timing out if the condition is not met.
   - This module does not support running commands in configuration mode.
-    Please use M(dnos9_config) to configure Dell OS9 devices.
+    Please use M(dellos9_config) to configure Dell OS9 devices.
 extends_documentation_fragment: dellos9
 options:
   commands:
     description:
-      - List of commands to send to the remote dnos9 device over the
+      - List of commands to send to the remote dellos9 device over the
         configured provider. The resulting output from the command
         is returned. If the I(wait_for) argument is provided, the
         module is not returned until the condition is satisfied or
@@ -77,25 +81,25 @@ vars:
 
 tasks:
   - name: run show version on remote devices
-    dnos9_command:
+    dellos9_command:
       commands: show version
       provider "{{ cli }}"
 
   - name: run show version and check to see if output contains OS9
-    dnos9_command:
+    dellos9_command:
       commands: show version
       wait_for: result[0] contains OS9
       provider "{{ cli }}"
 
   - name: run multiple commands on remote nodes
-    dnos9_command:
+    dellos9_command:
       commands:
         - show version
         - show interfaces
       provider "{{ cli }}"
 
   - name: run multiple commands and evalute the output
-    dnos9_command:
+    dellos9_command:
       commands:
         - show version
         - show interfaces
@@ -134,7 +138,7 @@ warnings:
 from ansible.module_utils.basic import get_exception
 from ansible.module_utils.netcli import CommandRunner, FailedConditionsError
 from ansible.module_utils.network import NetworkModule, NetworkError
-import ansible.module_utils.dnos9
+import ansible.module_utils.dellos9
 
 
 def to_lines(stdout):
@@ -169,9 +173,9 @@ def main():
                             'check mode, not executing `%s`' % cmd)
         else:
             if cmd.startswith('conf'):
-                module.fail_json(msg='dnos9_command does not support running '
+                module.fail_json(msg='dellos9_command does not support running '
                                      'config mode commands.  Please use '
-                                     'dnos9_config instead')
+                                     'dellos9_config instead')
             runner.add_command(cmd)
 
     for item in conditionals:
